+++ src/main/java/gregtech/api/graphs/consumers/NodeGTBaseMetaTile.java
@@ -4,14 +4,14 @@ import gregtech.api.interfaces.tileentity.IEnergyConnected;
 import gregtech.api.metatileentity.BaseMetaTileEntity;
 import java.util.ArrayList;
 
-//consumer for gt machines
+// consumer for gt machines
 public class NodeGTBaseMetaTile extends ConsumerNode {
     public NodeGTBaseMetaTile(int aNodeValue, BaseMetaTileEntity aTileEntity, byte aSide, ArrayList<ConsumerNode> aConsumers) {
         super(aNodeValue, aTileEntity, aSide, aConsumers);
     }
 
     @Override
-    public int injectEnergy(int aVoltage, int aMaxAmps) {
+    public int injectEnergy(long aVoltage, long aMaxAmps) {
         return (int)((IEnergyConnected) mTileEntity).injectEnergyUnits(mSide,aVoltage, aMaxAmps);
     }
